backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]>=2.0.35
alembic>=1.14.0
aiosqlite>=0.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
anthropic>=0.40.0
httpx>=0.28.0
python-multipart>=0.0.18
Pillow>=11.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
openai
